Fixing shadows not rendering on some hardware
Shadow commands were being given inconsistant function constants on compliation. The shader is looking for the source types to be properly encoded as constants, but the shadow shader was hard coded to look for the source type constant for layer 3. This caused layers 1 and 2 to be dropped and not properly bound because the shader assumed there was nothing there.