Remove special handling of skip/piks tokens and lines in `hsStream`
Previously, a pair of "skip" and "piks" tokens (for `GetToken`) or lines
(for `ReadLn`) was treated like a block comment - anything between them
was simply ignored. Because this was part of hsStream, it affected many
file types (.csv, .fni, .ini, .sdl, .srt), even though many of them
already have proper syntax for comments.

This feature (?) isn't documented anywhere and apparently no existing
code or data files rely on this, so just get rid of it.