Define `HS_DEBUGGING` directly instead of going through `_DEBUG`
This way, the same CMake code can be used for all compilers. Unlike
`_DEBUG`, which is automatically defined by MSVC, `HS_DEBUGGING` never
conflicts with compiler built-in macros.

The PhysX headers require `_DEBUG` to be defined for debug builds, even
when using a non-MSVC compiler, so the compiler-dependent definition of
`_DEBUG` is still needed there.