Maintain game window size when the DPI is changing.
Per Colin, on macoS, when the scale factor is changed, the game window
stays the same size. This uses a window message added in Windows 10
v1703 to ensure that the game window stays the same size when the DPI
changes. Unfortunately, at this time, Windows 8.1 - Windows 10 v1607
will still resize.