Fix non-standard uses of preprocessor paste operator ##
Pasting an identifier and a symbol is non-standard and unnecessary -
without the ## operator, the identifier is also expanded as expected.
MSVC accepts this use of ##, but CLion (apparently via clangd) reports
it as an error ("Pasting formed '...', an invalid preprocessing token").

According to the standard, ## inside a string literal has no effect -
the ## and any adjacent identifiers are taken literally and not expanded
or pasted. The correct standard solution is to use the stringification
operator # and (if needed) implicit string concatenation.