Use DirectXMath for matrix multiplication on Windows.
The benefit that we're seeing from our hand-rolled SIMD implementations
is waning on Windows due to improved auto-vectorization under MSVC.
Further, with the moves being made on the Apple side, it makes more
sense to delegate the optimizations to the OS where we can. DirectXMath
can be compiled to use FPU, SSE2, SSE4, AVX, AVX2, and/or NEON
instructions. Almost everyone playing on Windows has SSE2, and recent
versions of VS build with SSE2 instructions, even for 32-bit builds.
DirectXMath defaults to SSE2. It makes a lot of sense, then, to remove
the dynamic dispatch pointer chase and remove our manually maintained
optimization in favor of DirectXMath. Maybe someone enterprising could
add DirectXMath support for Linux in the future.