Remove non-functional plLocalization::UsingUnicode method
If this method returned true, plDynamicTextMap would attempt to load a
font named "Unicode", but no such font exists in the game data. The
method returned false for all languages except Japanese, which has no
actual translations and is not a user-selectable language, so this code
was practically dead already.

If Japanese (or another non-Latin language) is ever actually supported,
some sort of font fallback code will become necessary again. Ideally,
that code should examine the font and actual text to check whether a
fallback is necessary, instead of relying on an abstract flag whether a
language "uses Unicode" (which isn't even correct terminology).