Allow swimming forward to satisfy walking forward condition.
The `plObjectInVolumeAndFacingDetector` is used to, for example, trigger
ladder mounts and dismounts when walking forward. It makes sense that we
might want to do similar animations when swimming forward (eg climbing
up out of a body of water using a special animation). Unfortunately,
swimming is a special brain type with a forward movement behavior that
isn't working. So, update the logic to allow swimming forward to trigger
the walking forward condition.