Expose just the hsGMatState enums as a header
For shader-based pipelines with Metal and OpenGL, we need to make use of
these hsGMatState flag values in the shader code, but the shaders can't
include the existing hsGMatState.h file with C++ class definitions.

This extracts just the enum values into their own header that can be
imported by shaders, with one major caveat: Rather than plain enum
declarations, the enums are generated from macro statements.

You must define `GMAT_STATE_ENUM_START`, `GMAT_STATE_ENUM_VALUE`, and
`GMAT_STATE_ENUM_END` before importing this header.

This is because OpenGL shaders don't support enums except as an
extension, so it would be better for these to be defined as constants in
the shader in that situation.