Convert specific unicode ranges in `plFontConverter`.
Previously, the user was allowed to manually specify a number of
characters to convert. This is a bit of an odd choice and doesn't match
what the other (read: non-freetype) methods do (convert everything). We
probably don't want all of the glyphs in unicode fonts (for now), but we
do want the Latin and Cyrillic ones for sure.