Disallow implicit conversion of plKey to plKeyImp
This conversion now requires explicitly calling plKeyImp::GetFromKey,
making it easier to find which code really needs it. The main plKey.h
header no longer exposes plKeyImp in any way.

A lot of code implicitly used this conversion to check if a plKey's
underlying pointer is nullptr or not. This is now solved with an
implicit conversion to bool.