Specify older Pillow version to fix cairosvg build
Update to latest cairosvg too, since that seems to work fine.

On July 1st, a new major version of Pillow was released, and it requires
zlib to compile. For whatever reason, that's not working when pip tries
to install it. However cairosvg is fine using the older version so it's
easiest to rollback for the moment to fix the builds.

Co-Authored-By: Adam Johnson <AdamJohnso@gmail.com>