Add a final (upcasting) `ptGUIControl` getter to ptGUIDialog.
This allows the scripter to avoid indirections through `ptKey` in their
python code such that instead of writing something like:
```python
ctrl =
ptGUIControlButton(MyGUIAttr.dialog.getControlFromTag(kButtonTag))
ctrl.show()
```

The following can be used:
```python
ctrl = MyGUIAttr.dialog.getControlModFromTag(kButtonTag))
ctrl.show()
```

This allows for more concise Python code and avoids constructing
temporary `pyKey` objects in the engine. This API change is backwards
incompatible, so I opted to make this a new method. I elected to not
make `ptGUIDialog` a mapping type because controls can be fetched by
both ID and tag, which are distinct values. It's better to be explicit
in this case, IMO.