Update README to point to website instructions
It's better to maintain these in a single place rather than multiple
places where they can get out of date and out of sync (although yes it
would be nice for the instructions to be available after cloning the
repo).

Also update the list of libraries to keep it accurate.