Prefer to grab the server's SDL descriptors.
This removes the singlular `/SkipPreload` command line argument in favor
of two arguments: `/LocalPython` and `/LocalSDL`. `/LocalData` now
implies the former but not the latter. The reason for this change is
that the SDL is a contract between the server and the client about how
exactly an SDL blob is formatted. If SDL blobs don't have the same
format between the client and server, this can have fairly bad
implications, ranging from remote players being disconnected with an
error, the player with the wrong SDL getting an error, or the server
crashing.