Don't trust the local machine's system time.
Previously, when asking for the server's time, the calculation was done
by basically offsetting the local system time by some previously
calculated delta. This was basically useless because the user can easily
change the system time out from under us and cause huge deltas. The
attempt to fix things in #173 by using `WM_TIMECHANGE` was a good idea
but ultimately incorrect because `hsTimer` is a monotonic timer. At this
point, it uses `std::chrono::steady_clock` as its backend, so attempting
to use `hsTimer` to recalculate the server/client time offset is
useless.

Therefore, this changes us to simply save the last time value that we
received from the game server and the monotonic time that we received
that update. This is definitely not perfect because we don't consider
things like round trip time for the message and time spent in the update
loop. However, it *is* a clear improvement in that now the KI will
always display what the server thinks Mountain time is, and it should be
impossible to perform system clock based exploits (such as instantly
baking pellets)... without any platform specific code!