Allowing Github Actions to handle zipping
There be dragons here. We originally kept Github Actions from performing the zip because our archive had symlinks in it, which it turns into hard copes of files. Xcode's archive and code signing tooling expects those symlinks, so we weren't able to sign.

As of right now - there are no symlinks in the archive. So things should be ok. If any frameworks are added this will break again. Hopefully by that point Github has resolved their issues.