Don't expect specific value for AF_INET from system headers
Although AF_INET is always 2 on every relevant system, it's better not
to assume this. Here, we use AF_INET only as a marker value in a
serialization format and never pass it to any system functions, so it's
more important that the value is consistent between systems than that it
matches the system constant.