Fix plNetAddress::Read/Write on big-endian systems
fHost is always stored big-endian in memory, regardless of the host
endianness. It's thus incorrect to read/write it using
ReadLE32/WriteLE32, because then the byte order in the stream changes
depending on the host byte order!

On little-endian systems, fHost has always been read/written big-endian
(because ReadLE32/WriteLE32 don't swap anything on little-endian
systems). To make big-endian systems behave the same way, the field
needs to be read/written as-is with no endianness conversion.