Fix ahnyQuab ogg decompression crashes.
This issue was introduced when `hsStream::Close()` was removed in favor
of RAII. Unfortunately, the patcher is handing off files before it
deletes the patcher stream, which causes race conditions around zlib
decompression not being complete. We need to explicitly close the
underlying stream before performing any handoffs to prevent crashes.