Fix #1524
Xcode is a multi-config generator that may not define `NDEBUG` if
`CMAKE_BUILD_TYPE` is unset and the project is built in Release or
RelWithDebInfo via `cmake --build`. Therefore, be sure to define
`NDEBUG` if it's a non-debug config. Also, simplify the define for
`_DEBUG`. It's OK to double-define something.