Store NetMsg protocols/channels locally instead of in a global registry
The server side probably needs this so it can look up the right
protocol for each incoming connection, but the client always knows which
protocol it's using and so doesn't need a global registry.

This reduces the amount of bookkeeping needed in pnNetCli without
requiring much more work in plNetGameLib.