Add xEntryCam.py
This adds support for entry camera spawn points from Korman. The idea is
that, when spawning in, a specific camera will be pushed inside of a
specific region. This camera will only ever be pushed when spawning in.
Therefore, we need to disable all entry camera regions after linking in.