Fix #1508.
As it turns out, the responder state combo box has the owner draw style
applied. So, `ifdef`ing the draw method out causes the box to be blank.
Who would have thought it? While I was here, I cleaned up the colors to
not look out of place on Max 2022.