Add the ability to query DCM capability from Python.
Myst V didn't have this function either. That's interesting because, in
theory, we may be presented with a video card that doesn't support the
Direct3D caps needed to perform the effect. At this stage in the game,
everyone should support DCMs, so this is just an exercise in technical
correctness.