Cleanup key binding code.
Previously, it was possible to get the key map chronicle into an invalid
state with the same key bound to multiple controls. You can reproduce
that by binding UpArrow to "start chat" then rebinding UpArrow to "move
forward". The chronicle will then have two entries for UpArrow in it.
The bindings will behave as you expect until you quit and restart Uru,
at which point UpArrow will be bound to "start chat" again. The only way
to fix this is by resetting the key map to default.

To fix this, I cleaned up the code to make the ownership of the key map
more clear. Previously, the code half-handedly updated the vault
chronicle value and the key map. Now, we bind any input directly, and
completely rewrite the chronicle each time a key is bound. In this way,
the UI display, key bindings, and key chronicle all have a single source
of truth.

This also includes general code cleanups, modernizing the Python. For
example, we can now rely on the key dictionary being stored in the same
order that it's initialized.