Detect pthread_setname_np with CMake
In some non-standard setups that slip through our checks, this function
doesn't exist and we should guard against calling it. Since it's only
used for debugging, let's just use CMake to check if it's defined in the
header rather than complicated conditionals in the code for the 4
different possible versions of this that might exist.