Reducing required flags
Following the WoW client’s lead - going to allow files to inherit owner of the user running the application - but the file permissions should be maintained. Could manually 777 everything - but for now we’ll let the tarfile set that.

Technically this could cause issues as the permissions of what the installer wrote and what is on disk may diverge - but that won’t effect functionality - only “repair permissions” scans.