Use system wide file locks to synchronize log files.
The global semaphore has been observed to not be released on unix like
platforms when the process exits. So, we convert to file locks, which
will be released on process exit on unix like platforms. Unfortunately,
Win32 file locks and POSIX file locks have slightly different semantics
in that Win32 locks are mandatory and POSIX locks are advisory, which
can result in slightly different behavior. OTOH, this should fix the
client deadlocking until the system is restarted, so this should be an
overall win.