Allow building macOS client with Makefiles
This removes the strict requirement to use the Xcode project generator,
but still requires the Xcode build tools (and specifically the Metal
compiler) to be installed.

I expect that our official stance will always be that the Xcode project
generator is the recommended and supported way to do macOS builds, but
any day I don't have to launch Xcode is a good day.