Add plAIDestroyedMsg.
There is a long-standing bug on MOULa in the Ahnoying Quabs
implementation that causes the game to crash when the owner of the
VarSync game leaves. The problem is that Cyan's server despawns all
clones that a player spawns when they leave the Age. The Ahnoying quabs
script has no way to realize that, though, and tries to access a dead
critter brain, causing a crash. This broadcast will help us detect that.