Add plAIGoToGoalMsg.
This message is designed to be used to send updates to the critter's
goal position to remote clients. It is not currently doing so but will
in a future commit. The reason for this is how Cyan had the quabs
synchronizing across clients... In the old version of the quabs, only
the owner of the var sync game will simulate the movement of the quab.
Everyone else sees the quab bouncing around every second or so due to
SDL position updates. In the current H'uru script, everyone is
simulating the quab's movement, but the ownership model is a bit
unclear. Every client may set a different goal position based on where
they think all of the players are. This is an attempt to get the best of
both worlds - smoother movement due to everyone moving the quab to the
goal and improved ownership semantics.