Resolve Max texture paths before exporting
Max seems to save the original full file path to the texture image in
its data, but happily resolves it against the current working directory
when a file is opened in the editor. Except that it doesn't ever update
the stored full file path, so when the PlasmaMax plugin asks for files
it can get paths that don't exist. This is a problem because it tries
to detect changes to textures based on modified times, and nonexistent
files have nonexistent modified times, which can lead to the same image
getting exported multiple times because PlasmaMax tries to treat each
instance of it as a new texture.

The Max SDK includes a function to resolve texture file paths, so we'll
call that in a bunch of spots before we use the texture file path for
anything.