Work around not-found textures.
As discussed in IRC and Discord, 3ds Max does some manipulation and
pretty much always successfully loads the bitmap data. Unfortunately,
we are trying to access the file manually here, which might be an
absolute path to a file on another machine. This path may not be valid
on our system, which fails, causing textures to be duplicated.