Remove special case for plain IP addresses when connecting to servers
AsyncAddressLookupName can handle plain IP addresses just as well as
host names, so there's no need for this special case.

This fixes the client being unable to connect to server addresses that
contain a plain IP address with a port number (either in server.ini or
when returned by the gatekeeper server). A host name with a port number
has always worked fine.