Hopefully better cross-compiling support with Python
CMake historically has had issues with finding dependencies for the
target system and the host system in the same project. vcpkg has its own
concept of host dependencies, but that doesn't generalize to the rest of
CMake.

In Plasma's case, the only host dependency that is causing is pain is
the Python interpreter, to run the resources.dat building step on the
host at compile time.

CMake 4 introduces the idea of Python artifact prefixes, to allow
finding Python multiple times. They give cross-compiling and needing the
host interpreter as an example of where this is useful (although it's
unclear to me how exactly it's told to find the host interpreter vs the
target interpreter).

This change should allow us to opt-in to the prefixed behaviour on CMake
4, while retaining the existing behaviour on CMake 3 with a conditional
assignment to the prefixed executable variable.