Change .py license headers to comments and remove encoding declarations
The .py license header is now identical to the .sdl license header.

Previously, the license headers were interpreted as module docstrings,
causing issues for modules that contain (what was supposed to be) a real
docstring. In particular, it was impossible to access the real docstring
at runtime (__doc__ was set to the license header) and it broke
__future__ imports (which are only recognized when they come immediately
after the docstring - a "second docstring" in between is not allowed).

The explicit UTF-8 encoding declarations are no longer necessary,
because UTF-8 is always the default encoding in Python 3.