Remove ptAgeVault.getAgesIOwnFolder in favor of .getBookshelfFolder
The implementation of the two methods is identical, so this doesn't
change any behavior.

However, the name getBookshelfFolder more accurately describes the
method's behavior and avoids confusion with ptVault.getAgesIOwnFolder,
which does *not* do the same thing as the ptAgeVault method:

* ptVault.getAgesIOwnFolder can be called in any age and returns the age
  instances owned by the current avatar.
* ptAgeVault.getBookshelfFolder (formerly ptAgeVault.getAgesIOwnFolder)
  only succeeds when in an instance of Relto and returns the age
  instances owned by the avatar that owns the Relto instance (i. e. the
  books appearing in that Relto instance's bookshelf).

These two lists will be different if one avatar is visiting another
avatar's Relto!