Fix #1708.
Exclude regions were in practice not clearing because the PhysX CCT was
holding onto the collision so as long as the "details" of the collision
are largely the same. According to empirical evidence, once the
preFilter stage is passed, PhysX never questions an actor again unless
there's a position change or something of that nature. So, changing from
an exclude region group to a detector volume isn't enough. We actually
have to invalidate the character controller's cache to force
re-evaluation of the collision. This allows us to move through a
released exclude region immediately.

I suppose we could destroy and recreate the actor, but that feels like a
more expensive option than just telling the character controllers to
reevaluate their lives. Ideally, we would only do this for characters in
contact with an exclude region, but that starts to get complicated. It's
probably better to go with this simple fix.