Fix #1705.
The Nexus dialog never marks that it is closed. This causes the
open-dialog init to only run once. The dialog is fairly resilient and is
able to cope with that not happening, with the exception of the
Neighborhood button not being cleared out properly on exit. This fixes
the state tracking so that the button is reinitialized correctly.