Don't run the simulation while an Age is loading.
It feels like we used to do this, but I must have removed it during the
PhysX 4.1 rewrite. Why? It doesn't fix the "hitting everything" problem
during link-in, but it definitely helps. When linking into Garrison from
the bahro pole, you can now see yourself fall through the floor at the
main link in point!