Don't move from link out point to link in point.
Setting a CCT's position causes the PhysX character code to set the
kinematic target of the underlying actor. When the kinematic target is
set, on the next simulation step, PhysX will move the actor from the
previous location to the target, hitting and triggering everything along
the path. This is not desirable when dealing with linking to a new Age,
so just overwrite the global pose with the kinematic target that the CCT
wants upon link-in.