Fix #1716.
Fixes #1716 by making the implicit logic around focused explorers
explicit. Gone is the idea of "toggling" - now the state must be
explicitly requested. Also, apparently, there was some problem with a
default state returning that Cyan was hacking around. Instead, when the
GUI dialog is shown, we simply fixup everything.