Add link support to text in journals.
Images already support links by adding the `link` attribute and setting
it to an integer id. I have extended this into text links as well by
adding the `<link>` tag. To establish a link, set the `event` attribute
to any integer. To clear a link, set the `event` attribute to `clear`.
Closing tags, such as `</link>` don't exist in EsHTML, so I did not opt
to introduce a new pattern.

BUT WHAT ABOUT URLS??? I hear you! The current link support in the KI
uses Python's webbrowser module to open links. This module is a good
example of a cross platform way to open links, and it weighs in at
roughly 24 KB!!! No way am I reinventing that wheel. So, we would prefer
for Python to open links for us. With that in mind, we need the URLs in
Python. Unfortunately, book event notifications only send out integer
IDs. I could change the event to include a string, but that opens up a
whole can of compatibility worms. Instead, I opted to process the
journal contents in xJournalBookGUIPopup.py - you can use `<link
href="https://foo.com">` there! The python script will rewrite this to
an integer ID and convert that ID back to the requested URL on
notification. Sweet!