Rename global DebugMsg function to hsDebugPrintToTerminal
Fixes name collisions with multiple other, non-global functions/methods
also named DebugMsg, e. g. in plLoggable, plClient, and the SDL code.

Also adds a doc comment explaining the purpose of this function and why
it's usually not the right choice. The new function name is
intentionally long to make its purpose clear and discourage using it for
regular logging.