Remove incorrect unreachable declaration in DEFAULT_FATAL
In debug builds, the code after FATAL is in fact reachable, because the
failed assertion dialog offers the user the option to ignore the
failure.

hsDebugAssertionFailed is already declared as [[noreturn]] where
appropriate (i. e. in non-debug builds). This should be enough for
compilers to figure out when the code after the call is unreachable.