Remove unused and half-implemented ptDniInfoSource
Although PtGetAgeName claims to be "DEPRECIATED" in favor of
ptDniInfoSource, the latter is also marked as "DO NOT USE". There are
more than 100 uses of PtGetAgeName in the Python scripts and
exactly 0 uses of ptDniInfoSource, so let's remove the misleading
documentation and drop the unused code.

No functionality is lost by doing this. getAgeCoords always returned
None, and getAgeTime called an unimplemented method in the vault code.
Only getAgeName and getAgeGuid actually worked, but the same
functionality is provided by ptAgeVault, PtGetAgeInfo, and PtGetAgeName.