Optimizing more of the light management code
IDisableLightsForShadow can go, and only encoding the number of lights needed in the command buffer.

I’m removing the FIXMEs on adding dirty flags. I think the amount of overhead needed to maintain the flag is not better than doing a memory comparison. It all adds overhead but I’d rather have the smaller command buffers.