Only compile modules that can be imported.
plPythonPack was failing when it tried to compile modules that would not
import correctly. Also, don't try to compile modules that are part of
the frozen importlib. Bringing in the Python code for these modules
poisons the interpreter state, preventing modules that are imported
after this attempt to fail to eval.