Use variadic templates to construct pyEnums.
This means that we won't have to resize the list while we're populating
it - we know its length at compile time. The downside is that betweem
the `PYTHON_ENUM_START` and `PYTHON_ENUM_END` macro invocation is a
single function call. Right now, there is no code other than
`PYTHON_ENUM_ELEMENT`s between those, so, in practice, this isn't a big
deal.