Parse spawn point argument of Net.LinkWithOriginalBook properly
Previously, the argument was used directly as both the link title and
the spawn point. This is almost never correct.

In practice, people usually pass "Default:LinkInPointDefault:;" here.
That isn't a valid title or a valid spawn point, but because all code
involved is robust/lenient enough, it has the expected effect of linking
the player to the default spawn point and giving them a matching
linking book on Relto. As a side effect, it leaves some junk in the
corresponding Age Link vault node - the list of links will look like
"Default:LinkInPointDefault:;:Default:LinkInPointDefault:;:;" instead of
just "Default:LinkInPointDefault:;" as expected.

This commit fixes all of that by parsing the spawn point argument
properly, with error handling for invalid inputs.