Fixing Metal validation error around lighting
The Metal validator has started complaining that we’re passing a buffer that’s smaller than the argument size. This validation is tripped even if I change the argument in the shader to a pointer.

The idea was that I could pass a partial structure of lights to save on bandwidth costs - but Metal no longer seems to like that. I’ll have to pass the full array of lights whenever they are updated.