Refactored Metal Light Encoding
The Metal renderer was encoding all lights for each material pass. This was expensive. Refactored so that all lights are sent at the beginning of the render pass, and then only indices are sent with each material pass. This should result in a lot less data being sent per render and seems to save abotu 5-10% CPU.

Scale is also sent with each material pass in since this can be different per material.