xConsoleCamera
Add a fairly simple script to handle looking at a console-like object.
Like for the "Kadish Tolesa's Vault Console" for example, but without any puzzle element in the script.

Detailed setup instructions:

```
1. activator
The clickable which triggers looking at the console.

2. camera
The camera we switch to when "activator" is clicked.

3. respUsingConsole (optional)
A two-state responder triggered when starting or stopping to look at the console.
The first state is called when the avatar stops looking at the console.
The second state is called when the avatar starts looking at the console.

4. sdlQuitConsole (optional)
The name of an SDL variable which, when changed, forces any player to stop interacting with the console.
```

This script, from @Jrius, works on both POTS and Destiny.
It may need some ajustments to be accepted on H-uru, but like I'm not a code-dev, I'm counting on code-devs to tell me what to do, or to do the ajustments.